fix/delete tests
