+++ components/arm/client_test.go
@@ -22,6 +22,13 @@ import (
 	"go.viam.com/rdk/testutils/inject"
 )
 
+const (
+	testArmName    = "arm1"
+	testArmName2   = "arm2"
+	failArmName    = "arm3"
+	missingArmName = "arm4"
+)
+
 func TestClient(t *testing.T) {
 	logger := logging.NewTestLogger(t)
 	listener1, err := net.Listen("tcp", "localhost:0")
