+++ components/arm/fake/fake_test.go
@@ -5,11 +5,14 @@ import (
 	"math"
 	"testing"
 
+	"github.com/golang/geo/r3"
 	"go.viam.com/test"
 
+	"go.viam.com/rdk/components/arm"
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/referenceframe"
 	"go.viam.com/rdk/resource"
+	"go.viam.com/rdk/spatialmath"
 )
 
 func TestReconfigure(t *testing.T) {
@@ -130,4 +133,145 @@ func TestJointPositions(t *testing.T) {
 	inputs, err := arm.CurrentInputs(ctx)
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, sampleInputs, test.ShouldResemble, inputs)
+}
+
+const (
+	testArmName    = "arm1"
+	testArmName2   = "arm2"
+	failArmName    = "arm3"
+	missingArmName = "arm4"
+)
+
+func TestXArm6Locations(t *testing.T) {
+	// check the exact values/locations of arm geometries at a couple different poses
+	logger := logging.NewTestLogger(t)
+	cfg := resource.Config{
+		Name:  arm.API.String(),
+		Model: resource.DefaultModelFamily.WithModel("fake"),
+		ConvertedAttributes: &Config{
+			ModelFilePath: "../example_kinematics/xarm6_kinematics_test.json",
+		},
+	}
+
+	notReal, err := NewArm(context.Background(), nil, cfg, logger)
+	test.That(t, err, test.ShouldBeNil)
+
+	t.Run("home location check", func(t *testing.T) {
+		checkMap := make(map[string]r3.Vector)
+		checkMap["rdk:component:arm:base_top"] = r3.Vector{
+			0.000000000000000000000000,
+			0.000000000000000000000000,
+			160.000000000000000000000000,
+		}
+		checkMap["rdk:component:arm:upper_arm"] = r3.Vector{
+			0.000000000000000000000000,
+			0.000000000000000000000000,
+			402.000000000000000000000000,
+		}
+		checkMap["rdk:component:arm:upper_forearm"] = r3.Vector{
+			102.997474683058328537299531,
+			0.000000000000000000000000,
+			502.002525316941671462700469,
+		}
+		checkMap["rdk:component:arm:lower_forearm"] = r3.Vector{
+			131.000000000000000000000000,
+			-27.500000000000000000000000,
+			274.200000000000000000000000,
+		}
+		checkMap["rdk:component:arm:wrist_link"] = r3.Vector{
+			206.000000000000000000000000,
+			10.000000000000000000000000,
+			141.500000000000000000000000,
+		}
+
+		in := make([]referenceframe.Input, 6)
+		geoms, err := notReal.ModelFrame().Geometries(in)
+		test.That(t, err, test.ShouldBeNil)
+		geomMap := geoms.Geometries()
+		b := locationCheckTestHelper(geomMap, checkMap)
+		test.That(t, b, test.ShouldBeTrue)
+	})
+	//nolint:dupl
+	t.Run("location check1", func(t *testing.T) {
+		checkMap := make(map[string]r3.Vector)
+		checkMap["rdk:component:arm:base_top"] = r3.Vector{
+			0.000000000000000000000000,
+			0.000000000000000000000000,
+			160.000000000000000000000000,
+		}
+		checkMap["rdk:component:arm:upper_arm"] = r3.Vector{
+			-13.477511247321800169629569,
+			0.000000000000000000000000,
+			401.325562312533520525903441,
+		}
+		checkMap["rdk:component:arm:upper_forearm"] = r3.Vector{
+			83.174566602088916056345624,
+			0.000000000000000000000000,
+			516.742582359123957758129109,
+		}
+		checkMap["rdk:component:arm:lower_forearm"] = r3.Vector{
+			158.566974381217988820935716,
+			-27.362614545145717670493468,
+			299.589614460540474283334333,
+		}
+		checkMap["rdk:component:arm:wrist_link"] = r3.Vector{
+			259.050559200277916716004256,
+			18.072894555453149934010071,
+			192.543551746158527748775668,
+		}
+
+		in := []referenceframe.Input{{Value: 0}, {Value: -0.1}, {Value: -0.1}, {Value: -0.1}, {Value: -0.1}, {Value: -0.1}}
+		geoms, err := notReal.ModelFrame().Geometries(in)
+		test.That(t, err, test.ShouldBeNil)
+		geomMap := geoms.Geometries()
+		b := locationCheckTestHelper(geomMap, checkMap)
+		test.That(t, b, test.ShouldBeTrue)
+	})
+	//nolint:dupl
+	t.Run("location check2", func(t *testing.T) {
+		checkMap := make(map[string]r3.Vector)
+		checkMap["rdk:component:arm:base_top"] = r3.Vector{
+			0.000000000000000000000000,
+			0.000000000000000000000000,
+			160.000000000000000000000000,
+		}
+		checkMap["rdk:component:arm:upper_arm"] = r3.Vector{
+			-26.820359657333263214695762,
+			0.000000000000000000000000,
+			399.308988008567553151806351,
+		}
+		checkMap["rdk:component:arm:upper_forearm"] = r3.Vector{
+			60.777555075062835499011271,
+			0.000000000000000000000000,
+			530.142781900699674224597402,
+		}
+		checkMap["rdk:component:arm:lower_forearm"] = r3.Vector{
+			180.312201371473605604478507,
+			-26.951830890634148829576588,
+			333.355009225598280409030849,
+		}
+		checkMap["rdk:component:arm:wrist_link"] = r3.Vector{
+			297.258065257027055849903263,
+			27.068045067389423508075197,
+			256.363984524505951867467957,
+		}
+
+		in := []referenceframe.Input{{Value: 0}, {Value: -0.2}, {Value: -0.2}, {Value: -0.2}, {Value: -0.2}, {Value: -0.2}}
+		geoms, err := notReal.ModelFrame().Geometries(in)
+		test.That(t, err, test.ShouldBeNil)
+		geomMap := geoms.Geometries()
+		b := locationCheckTestHelper(geomMap, checkMap)
+		test.That(t, b, test.ShouldBeTrue)
+	})
+}
+
+func locationCheckTestHelper(geomList []spatialmath.Geometry, checkMap map[string]r3.Vector) bool {
+	for _, g := range geomList {
+		vecCheck := checkMap[g.Label()]
+		vecActual := g.Pose().Point()
+		if !spatialmath.R3VectorAlmostEqual(vecCheck, vecActual, 1e-2) {
+			return false
+		}
+	}
+	return true
 }
